Move Puzzle Page validation into its own file (PR review feedback)
Per PR #16700 review feedback ('wondering if it might be better placed in
something like puzzles.validate.ts... so we're not mixing too much code
with the user-related logic'), moved isPuzzlePageInstance/
validateAsPuzzlePageType out of the general src/model/validate.ts into a
new dedicated src/model/validate.puzzlePage.ts.

- src/model/validate.ts: exported the small set of generic helpers Puzzle
  Page validation needs (isRecord, isNonEmptyString, isPuzzlesConfig,
  isPuzzleItem, editions) so the new file can reuse them rather than
  duplicating logic. Removed isPuzzlePageInstance/validateAsPuzzlePageType
  and the now-unused FEPuzzlePageType import.
- src/model/validate.puzzlePage.ts (new): exports validateAsPuzzlePageType,
  built on top of the exported helpers above.
- src/server/handler.puzzlePage.web.ts: import updated to the new file.
- src/model/validate.puzzlePage.test.ts: import updated to the new file
  (test contents/behaviour unchanged).

Note on the 'mirror the existing pattern' framing: I checked first, and
there was actually no pre-existing validate.<pageType>.ts-style file for
any other page type in this repo (including the unrelated Puzzles Hub) -
every other page type's validator, including validateAsPuzzlesPageType,
still lives in the shared validate.ts; only their *tests* are split into
per-page-type files. This commit establishes a new, more separated
convention for Puzzle Page specifically, per this review comment, rather
than genuinely mirroring an existing one - flagging this so the
distinction is clear for future reference.

tsc --noEmit clean, eslint clean, 8 suites / 82 tests passing (Puzzle
Page + Puzzles Hub + general validate, no regressions).
